% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_make_template.R
\name{cc_make_template}
\alias{cc_make_template}
\title{Make a correction template file}
\usage{
cc_make_template(path = "correction_template.csv")
}
\arguments{
\item{path}{path for the correction template file}
}
\description{
Make a correction template file
}
